+++ examples/infinite_list_state_machine/lib/posts/view/posts_list.dart
@@ -1,0 +1,64 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:infinite_list_state_machine/posts/posts.dart';
+
+class PostsList extends StatefulWidget {
+  @override
+  _PostsListState createState() => _PostsListState();
+}
+
+class _PostsListState extends State<PostsList> {
+  final _scrollController = ScrollController();
+
+  @override
+  void initState() {
+    super.initState();
+    _scrollController.addListener(_onScroll);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<PostBloc, PostState>(
+      builder: (context, state) {
+        if (state is PostInitial)
+          return const Center(child: CircularProgressIndicator());
+
+        if (state is PostError)
+          return const Center(child: Text('failed to fetch posts'));
+
+        if (state.posts.isEmpty) return const Center(child: Text('no posts'));
+
+        return ListView.builder(
+          itemBuilder: (BuildContext context, int index) {
+            return index >= state.posts.length
+                ? BottomLoader()
+                : PostListItem(post: state.posts[index]);
+          },
+          itemCount: state is PostEndReached
+              ? state.posts.length
+              : state.posts.length + 1,
+          controller: _scrollController,
+        );
+      },
+    );
+  }
+
+  @override
+  void dispose() {
+    _scrollController
+      ..removeListener(_onScroll)
+      ..dispose();
+    super.dispose();
+  }
+
+  void _onScroll() {
+    if (_isBottom) context.read<PostBloc>().add(PostFetchRequested());
+  }
+
+  bool get _isBottom {
+    if (!_scrollController.hasClients) return false;
+    final maxScroll = _scrollController.position.maxScrollExtent;
+    final currentScroll = _scrollController.offset;
+    return currentScroll >= (maxScroll * 0.9);
+  }
+}