+++ examples/infinite_list_state_machine/test/app_test.dart
@@ -1,0 +1,13 @@
+import 'package:infinite_list_state_machine/app.dart';
+import 'package:infinite_list_state_machine/posts/posts.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  group('App', () {
+    testWidgets('renders PostsPage', (tester) async {
+      await tester.pumpWidget(App());
+      await tester.pumpAndSettle();
+      expect(find.byType(PostsPage), findsOneWidget);
+    });
+  });
+}