+++ examples/infinite_list_state_machine/test/posts/models/post_test.dart
@@ -1,0 +1,14 @@
+// ignore_for_file: prefer_const_constructors
+import 'package:infinite_list_state_machine/posts/models/models.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  group('Post', () {
+    test('supports value comparison', () {
+      expect(
+        Post(id: 1, title: 'post title', body: 'post body'),
+        Post(id: 1, title: 'post title', body: 'post body'),
+      );
+    });
+  });
+}