+++ examples/infinite_list_state_machine/test/posts/view/posts_page_test.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:infinite_list_state_machine/posts/posts.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  group('PostsPage', () {
+    testWidgets('renders PostList', (tester) async {
+      await tester.pumpWidget(MaterialApp(home: PostsPage()));
+      await tester.pumpAndSettle();
+      expect(find.byType(PostsList), findsOneWidget);
+    });
+  });
+}